requirements.txt
streamlit==1.35.0
sounddevice==0.4.6
numpy==1.26.4
requests==2.31.0
python-dotenv==1.0.1
openai==1.30.1
faiss-cpu==1.8.0
langchain==0.2.1
pypdf==4.1.0
pandas==2.2.2
tiktoken==0.6.0
markdown==3.6
scikit-learn==1.5.0